Move `core/utils.py` into `core/util/vpk.py`
