+++ core/handlers/skybox_handler.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from core.folder_setup import folder_setup
 from core.handlers.file_handler import FileHandler
 from valve_parsers import VPKFile
-from core.utils import get_vpk_name
+from core.util.vpk import get_vpk_name
 
 
 def is_skybox_vmt(file_path: Path) -> bool:
