+++ core/util/vpk.py
